refactor:items.pyが持つ情報に地域グループ番号や月を追加
データベースに登録するための判別情報をわかりやすくする
